## Résumé
- Ajoute PHPStan (niveau 5) à la CI : analyse statique de `src/`,
`front/`, `hook.php`, `setup.php`, avec un bootstrap dédié qui charge
l'autoload GLPI sans exiger de connexion DB.
- Ajoute Trivy à la CI, restreint aux secrets/mauvaises configurations
(`--scanners secret,misconfig`) — le scan de vulnérabilités des
dépendances est déjà couvert par `composer audit`/`npm audit`, pas
dupliqué ici.
- `enlightn/security-checker` évalué comme "PHP Security Checker" dédié
puis écarté : sa dépendance `symfony/console` casse `plugin:activate`
une fois mergée dans le `vendor/` partagé avec le cœur GLPI (documenté
dans TROUBLESHOOTING.md).
- Corrige un vrai bug de documentation trouvé par PHPStan :
`SignatureStamper::apply()` retournait `signed_at` sans le déclarer dans
son `@return`.
- Deux corrections mineures associées (cast `(string)` sur un champ de
notification, cast `(int)` inutile sur un paramètre `Migration` qui
attend une chaîne).

## Test plan
- [x] `vendor/bin/phpstan analyse` propre (niveau 5) sur le code copié
dans un conteneur `glpi/glpi:11.0.8` frais
- [x] Suite PHPUnit complète (91 tests, 179 assertions) au vert après
les correctifs
- [x] `plugin:install`/`plugin:activate` fonctionnels (confirmé le
conflit `enlightn/security-checker` puis sa résolution)
- [x] Trivy testé en local (`--scanners secret,misconfig`), aucun
secret/mauvaise config détecté
- [x] CI verte sur `dev` (run 30753546290), les deux nouvelles étapes
incluses

+++ hook.php
@@ -142,7 +142,7 @@ function plugin_remise_getDropdown(): array {
 // ----------------------------------------------------------------------------------
 
 function plugin_remise_install(): bool {
-    $migration = new Migration((int) str_replace('.', '', PLUGIN_REMISE_VERSION));
+    $migration = new Migration(str_replace('.', '', PLUGIN_REMISE_VERSION));
 
     Config::install($migration);
     Template::install($migration);
@@ -214,7 +214,7 @@ function plugin_remise_install(): bool {
 }
 
 function plugin_remise_uninstall(): bool {
-    $migration = new Migration((int) str_replace('.', '', PLUGIN_REMISE_VERSION));
+    $migration = new Migration(str_replace('.', '', PLUGIN_REMISE_VERSION));
 
    foreach ([
         'glpi_plugin_remise_reminders',
